requirements.txt
astroid==2.4.0
certifi==2020.4.5.1
chardet==3.0.4
cycler==0.10.0
idna==2.9
isort==4.3.21
jieba==0.42.1
kiwisolver==1.2.0
lazy-object-proxy==1.4.3
matplotlib==3.2.1
mccabe==0.6.1
numpy==1.18.4
Pillow==7.1.2
pylint==2.5.0
pyparsing==2.4.7
python-dateutil==2.8.1
requests==2.23.0
six==1.14.0
toml==0.10.0
typed-ast==1.4.1
urllib3==1.25.9
wordcloud==1.7.0
wrapt==1.11.2
